--- src/pages/PremioBomDesempenho.js
@@ -1,9 +1,0 @@
-import Premio from "../components/Premio"
-
-const PremioBomDesempenho = () => {
-  return (
-    <Premio nota="alto"/>
-  )
-}
-
-export default PremioBomDesempenho;
