--- src/pages/PremioMedioDesempenho.js
@@ -1,9 +1,0 @@
-import Premio from "../components/Premio"
-
-const PremioMedioDesempenho = () => {
-  return (
-    <Premio nota="medio"/>
-  )
-}
-
-export default PremioMedioDesempenho;
