--- src/pages/PremioParticipacao.js
@@ -1,9 +1,0 @@
-import Premio from "../components/Premio"
-
-const PremioParticipacao = () => {
-  return (
-    <Premio nota="baixo"/>
-  )
-}
-
-export default PremioParticipacao;
